Crate problem098

Problem 98 (Anagramic squares)

Solution detail

This is not a conceptually difficult problem, just one that requires a careful implementation. The following process is one way of approaching it:

For each word in the list, sort its letters into increasing order and use that as a key into a HashMap to store this word (and all its anagrams).

Cast out all keys that only contain one word (they have no anagrams, so are useless).

Do the same for the string representations of all square numbers having length at most the length of the longest anagram word.

For each key (in both maps) compute the 'signature' of the word - the counts of how many times each letter appears, sorted in decreasing order. Also store all the words and squares in maps keyed by their signatures.

For each word signature, and for each word key and square key having that signature, loop over the words and squares that have those keys. If a mapping is found between the letters of the word and the digits of the square, then for each word which is an anagram of the given word, turn it into a number using the same correspondence, and if that number is square, bingo!

Structs

SignaturedHashMaps
A struct that represents anagrams from a given set of words, keyed in two ways. First, by the set of letters appearing in each word, and second, by the 'signature' of each word, where the signature is the counts of the distinct letters in a word, in decreasing order.

Functions

alphabetical_string
Sort the letters of the given word into alphabetical order.

answer
Solve the problem, returning the answer as a String

main

mapping
Find, if it exists, a mapping between the two given words, assumed to be of the same length

signature
Get the signature of the given word.

solve
Find the largest square number which is part of an anagramic square pair from the given words.